+++ utils/banner.js
@@ -8,7 +8,7 @@ function banner() {
   return [
     `Lime version ${Package.version}`,
     Package.description,
-    `2021 ${Package.author}, ${Package.license} license`,
+    `2021-2022 ${Package.author}, ${Package.license} license`,
     Package.homepage.slice(0, -7),
   ].join('\n');
 }